@@ -1,10 +1,11 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import styles from './Page.scss';
 
 export default class Feature extends React.Component {
   static propTypes = {
-    children: React.PropTypes.element.isRequired,
-    heading: React.PropTypes.string.isRequired
+    children: PropTypes.element.isRequired,
+    heading: PropTypes.string.isRequired
   };
 
   render() {
