@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import ReactDOM from 'react-dom';
 
 import AddBookmarkMutation from 'mutations/AddBookmarkMutation';
@@ -273,9 +274,9 @@ class BookmarksWidget extends React.Component {
 }
 
 BookmarksWidget.propTypes = {
-  widget: React.PropTypes.object.isRequired,
-  user: React.PropTypes.object.isRequired,
-  popoverWidgetVisibilityChanged: React.PropTypes.func.isRequired,
+  widget: PropTypes.object.isRequired,
+  user: PropTypes.object.isRequired,
+  popoverWidgetVisibilityChanged: PropTypes.func.isRequired,
 };
 
 export default BookmarksWidget;
