@@ -1,5 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import PropTypes from 'prop-types';
 
 import UpdateWidgetDataMutation from 'mutations/UpdateWidgetDataMutation';
 
@@ -317,9 +318,9 @@ class TodosWidget extends React.Component {
 }
 
 TodosWidget.propTypes = {
-  widget: React.PropTypes.object.isRequired,
-  user: React.PropTypes.object.isRequired,
-  popoverWidgetVisibilityChanged: React.PropTypes.func.isRequired
+  widget: PropTypes.object.isRequired,
+  user: PropTypes.object.isRequired,
+  popoverWidgetVisibilityChanged: PropTypes.func.isRequired
 };
 
 export default TodosWidget;
